apply some suggestions
