+++ signing/transaction-manager.js
@@ -138,27 +138,3 @@ export async function pollTxInclusion(txHash, iteration = 0) {
     )
   }
 }
-
-// limitation of the Ledger Nano S, so we pick the top 5 rewards and inform the user.
-export function getTop5RewardsValidators(rewards) {
-  const rewardsPerValidatorObject = rewards.reduce((all, reward) => {
-    return {
-      ...all,
-      [reward.validator.operatorAddress]:
-        Number(reward.amount) +
-        (Number(all[reward.validator.operatorAddress]) || 0),
-    }
-  }, {})
-  const rewardsPerValidatorAddresses = Object.keys(rewardsPerValidatorObject)
-  const rewardsPerValidatorArray = []
-  rewardsPerValidatorAddresses.forEach((validatorAddress, index) => {
-    rewardsPerValidatorArray.push({
-      validator: validatorAddress,
-      totalRewardAmount: Object.values(rewardsPerValidatorObject)[index],
-    })
-  })
-  return rewardsPerValidatorArray
-    .sort((a, b) => b.totalRewardAmount - a.totalRewardAmount)
-    .slice(0, 5)
-    .map((rewardPerValidator) => rewardPerValidator.validator)
-}
